requirements.txt
flask
pymongo[srv]
Flask-PyMongo
python-dotenv
flask-cors
flask-jwt-extended
